Handle db.properties with build system instead of git
This patch makes it easier to use different settings in
db.properties. Before, it was always necessary to stash changes before
pulling updates.

The new behaviour is to create the file with default values if it does
not exist. If it does exist, it remains untouched.
Also, the file is no longer managed by git, which prevents connection
credentials to be exposed.